+++ project3/src/main/java/com/example/project3/repositories/CarServiceTypeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CarServiceTypeRepository extends JpaRepository<CarServiceType, Long> {
+public interface CarServiceTypeRepository extends JpaRepository<CarServiceType, Integer> {
 }
